Support streaming .fw files into fwup
This was always intended to work, but at some point I added code to
check file sizes before reading them all the way. This doesn't work when
streaming. This commit also adds unit tests so hopefully this won't
break in the future.

+++ src/fatfs.c
@@ -302,7 +302,7 @@ DRESULT disk_read(BYTE pdrv,		/* Physical drive nmuber (0..) */
     if (amount_read < 0)
         amount_read = 0;
 
-    if ((off_t) amount_read != byte_count)
+    if ((size_t) amount_read != byte_count)
         memset(&buff[amount_read], 0, byte_count - amount_read);
 
     return 0;
